move action-related code into pyramid.config.actions
